requirements.txt
fastapi
uvicorn
PyPDF2
langchain
langchain-openai
pinecone>=3.0.0
openai
python-dotenv
requests
python-multipart
duckduckgo_search

